Add files via upload
The system allows clients to search and order food. The admin can see the orders when they login via the admin dashboard, and can manage the orders by modifying or deleting. The admin can also add food staffs to the user interface via the admin page and can update the details of existing foods (pictures) on the user interface.